Dev express

+++ src/App/Shared/GoldEx.Shared/DTOs/Products/UpdateGemStoneRequest.cs
@@ -1,3 +1,3 @@
 ﻿namespace GoldEx.Shared.DTOs.Products;
 
-public record UpdateGemStoneRequest(string Code, string Type, string Color, string? Cut, double Carat, string? Purity);
+public record UpdateGemStoneRequest(string Code, string Type, string Color, string? Cut, decimal Carat, string? Purity);
